# Conflicts:
#	domain/src/main/java/com/devsu/fintech/domain/ports/output/AccountRepositorySPI.java

+++ application/src/main/java/com/devsu/fintech/application/port/input/HasOpenedAccountsInputPort.java
@@ -2,5 +2,5 @@ package com.devsu.fintech.application.port.input;
 
 public interface HasOpenedAccountsInputPort {
 
-    boolean execute(Integer clientId);
+    boolean execute(Long clientId);
 }
