+++ infrastructure/src/main/java/com/devsu/fintech/infrastructure/adapter/messaging/dto/CheckAccountRequestDTO.java
@@ -1,4 +1,4 @@
 package com.devsu.fintech.infrastructure.adapter.messaging.dto;
 
-public record CheckAccountRequestDTO(Integer clientId) {
+public record CheckAccountRequestDTO(Long clientId) {
 }
